+++ addons/yat/src/YAT.cs
@@ -7,11 +7,8 @@ namespace YAT;
 
 public partial class YAT : Node
 {
-    [Signal] public delegate void YatReadyEventHandler();
-
 #nullable disable
     public Node Windows { get; private set; }
-    public BaseTerminal CurrentTerminal { get; set; }
 
     public YatEnable YatEnable { get; private set; }
     public DebugScreen DebugScreen { get; private set; }
@@ -26,16 +23,8 @@ public partial class YAT : Node
         YatEnable = GetNode<YatEnable>("./YatEnable");
         DebugScreen = GetNode<DebugScreen>("./Windows/DebugScreen");
         Commands = GetNode<RegisteredCommands>("./RegisteredCommands");
-        PreferencesManager = GetNode<PreferencesManager>("%PreferencesManager");
-
         TerminalManager = GetNode<TerminalManager>("./TerminalManager");
-        TerminalManager.GameTerminal.Ready += () =>
-        {
-            TerminalManager.GameTerminal.TerminalSwitcher.CurrentTerminalChanged +=
-            (BaseTerminal terminal) => CurrentTerminal = terminal;
-
-            _ = EmitSignal(SignalName.YatReady);
-        };
+        PreferencesManager = GetNode<PreferencesManager>("%PreferencesManager");
 
         Keybindings.LoadDefaultActions();
     }